fix(macros): Support variable-length delay encoding and add macro parser
- Change set_macro() events from (u8, bool, u8) to (u8, bool, u16)
  to support delays > 127ms using the wire protocol's variable-length
  encoding: short (2-byte) for 0-127ms, long (4-byte) for 128+ms
- Change set_text_macro() delay_ms from u8 to u16
- Add MacroEvent struct and parse_macro_events() for decoding raw
  macro data from the device using the same variable-length format
- Update sync wrappers to match new signatures

+++ iot_driver_linux/monsgeek-keyboard/src/sync.rs
@@ -341,7 +341,7 @@ impl SyncKeyboard {
     pub fn set_macro(
         &self,
         macro_index: u8,
-        events: &[(u8, bool, u8)],
+        events: &[(u8, bool, u16)],
         repeat_count: u16,
     ) -> Result<(), KeyboardError> {
         block_on(self.inner.set_macro(macro_index, events, repeat_count))
@@ -352,7 +352,7 @@ impl SyncKeyboard {
         &self,
         macro_index: u8,
         text: &str,
-        delay_ms: u8,
+        delay_ms: u16,
         repeat: u16,
     ) -> Result<(), KeyboardError> {
         block_on(
